backend/requirements.txt
annotated-types>=0.7.0
anyio>=4.9.0
certifi>=2025.4.26
charset-normalizer>=3.4.2
click>=8.1.8
colorama>=0.4.6
contourpy>=1.3.2
cycler>=0.12.1
exceptiongroup>=1.2.2
fastapi>=0.115.12
filelock>=3.18.0
fonttools>=4.57.0
fsspec>=2025.3.2
h11>=0.16.0
httpx>=0.27.0
idna>=3.10
Jinja2>=3.1.6
kiwisolver>=1.4.8
lap>=0.5.12
MarkupSafe>=3.0.2
matplotlib>=3.10.1
mpmath>=1.3.0
networkx>=3.4.2
numpy>=2.2.5
opencv-python-headless>=4.10.0.84
packaging>=25.0
pandas>=2.2.3
paramiko>=3.4.0
pillow>=11.2.1
psutil>=7.0.0
psycopg2-binary>=2.9.9
py-cpuinfo>=9.0.0
pydantic>=2.11.4
pydantic_core>=2.33.2
pyparsing>=3.2.3
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-multipart>=0.0.20
pytz>=2025.2
PyYAML>=6.0.2
requests>=2.32.3
scipy>=1.15.2
seaborn>=0.13.2
six>=1.17.0
sniffio>=1.3.1
starlette>=0.46.2
sympy>=1.14.0
torch>=2.7.0
torchvision>=0.22.0
tqdm>=4.67.1
typing-inspection>=0.4.0
typing_extensions>=4.13.2
tzdata>=2025.2
ultralytics>=8.3.124
ultralytics-thop>=2.0.14
urllib3>=2.4.0
uvicorn>=0.34.2
